+++ R/template_read.R
@@ -15,5 +15,21 @@
 #' @importFrom utils write.csv2 read.csv2
 
 template_read <- function(file){
+
   cras_table <- read.csv2(file, check.names = FALSE)
+
+  if(!("Authors" %in% names(cras_table)))
+    stop("A column named `Authors` is missing")
+
+  if(nrow(cras_table) < 1)
+    stop("The cras_table has zero rows")
+
+  if(!is.character(cras_table$Authors))
+    warning("Authors column is not of type character")
+
+  if(!all(vapply(cras_table[-1], is.numeric, FALSE)))
+    warning("Roles are not numeric, it can lead to unexpected behaviour")
+
+  return(cras_table)
+
 }